changed to double

+++ MNE/fiff/fiff_ch_info.h
@@ -116,9 +116,9 @@ public:
 
     fiff_int_t coil_type;       /**< Which kind of coil. */
 
-    Matrix<float,12,1, DontAlign>   loc;
-    Matrix<float,4,4, DontAlign>    coil_trans;  /**< Channel location */
-    Matrix<float,3,2, DontAlign>    eeg_loc;
+    Matrix<double,12,1, DontAlign>   loc;
+    Matrix<double,4,4, DontAlign>    coil_trans;  /**< Channel location */
+    Matrix<double,3,2, DontAlign>    eeg_loc;
     fiff_int_t                      coord_frame;
 
     //    fiff_ch_pos_t chpos;        /**< Channel location 15*/
